requirements.txt
annotated-types==0.7.0
asttokens==3.0.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.3.1
colorama==0.4.6
comm==0.2.3
contourpy==1.3.2
cycler==0.12.1
debugpy==1.8.15
decorator==5.2.1
deepinv==0.3.3
einops==0.8.1
exceptiongroup==1.3.0
executing==2.2.0
filelock==3.18.0
fonttools==4.58.5
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.45
h5py==3.14.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
ipykernel==6.30.1
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mpmath==1.3.0
natsort==8.4.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.6
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
parso==0.8.4
pexpect==4.9.0
pickleshare==0.7.5
pillow==11.3.0
platformdirs==4.3.8
prompt_toolkit==3.0.51
protobuf==6.33.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.2
pyzmq==27.1.0
requests==2.32.4
scikit-image==0.25.2
scikit-learn==1.7.2
scipy==1.15.3
sentry-sdk==2.46.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.5.10
torch==2.9.1
torch-dct==0.1.6
torchmetrics==1.7.3
torchvision==0.24.1
tornado==6.5.1
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
typing-inspection==0.4.2
typing_extensions==4.14.1
tzdata==2026.1
urllib3==2.5.0
wandb==0.23.0
wcwidth==0.2.13
zipp==3.23.0
